Trait re_space_view::data_query::DataQuery

pub trait DataQuery {
    // Required method
    fn execute_query(
        &self,
        ctx: &StoreContext<'_>,
        visualizable_entities_for_visualizer_systems: &PerVisualizer<VisualizableEntities>
    ) -> DataQueryResult;
}

The common trait implemented for data queries

Both interfaces return re_viewer_context::DataResults, which are self-contained description of the data to be added to a SpaceView including both the re_log_types::EntityPath and context for any overrides.

Required Methods

fn execute_query( &self, ctx: &StoreContext<'_>, visualizable_entities_for_visualizer_systems: &PerVisualizer<VisualizableEntities> ) -> DataQueryResult

Execute a full query, returning a DataResultTree containing all results.

auto_properties is a map containing any heuristic-derived auto properties for the given SpaceView.

This is used when building up the contents for a SpaceView.