Function re_space_view::external::re_chunk_store::external::arrow2::array::get_display

pub fn get_display<'a, F>(
    array: &'a (dyn Array + 'static),
    null: &'static str,
) -> Box<dyn Fn(&mut F, usize) -> Result<(), Error> + 'a>
where
    F: Write + 'a,

Returns a function that writes the element of array at position index to a Write, writing null to the null slots.