Function re_space_view::external::re_chunk_store::external::arrow2::compute::arity::unary_checked

pub fn unary_checked<I, F, O>(
    array: &PrimitiveArray<I>,
    op: F,
    data_type: DataType,
) -> PrimitiveArray<O>
where
    I: NativeType,
    O: NativeType,
    F: Fn(I) -> Option<O>,

Version of unary that creates a mutable bitmap that is used to keep track of checked operations. The resulting bitmap is compared with the array bitmap to create the final validity array.