Crate re_space_view::external::re_chunk_store::external::re_chunk::external::crossbeam::queue

Concurrent queues.

This crate provides concurrent queues that can be shared among threads:

ArrayQueue, a bounded MPMC queue that allocates a fixed-capacity buffer on construction.

SegQueue, an unbounded MPMC queue that allocates small buffers, segments, on demand.

Structs

ArrayQueue

A bounded multi-producer multi-consumer queue.

SegQueue

An unbounded multi-producer multi-consumer queue.