Module re_space_view::external::re_chunk_store::external::re_chunk::util

Functions

arrays_to_list_array

Create a sparse list-array out of an array of arrays.

arrays_to_list_array_opt

Create a sparse list-array out of an array of arrays.

filter_array

Applies a filter kernel to the given array.

pad_list_array_back

Create a new ListArray of target length by appending null values to its back.

pad_list_array_front

Create a new ListArray of target length by appending null values to its front.

sparse_list_array_to_dense_list_array

Given a sparse ArrowListArray (i.e. an array with a validity bitmap that contains at least one falsy value), returns a dense ArrowListArray that only contains the non-null values from the original list.