Crate re_space_view::external::re_query2

Provide query-centric access to the re_data_store.

Modules

clamped_zip

range_zip

Structs

ClampedZip1x1

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x2

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x3

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x4

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x5

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x6

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x7

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x8

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip1x9

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x1

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x2

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x3

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x4

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x5

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x6

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x7

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x8

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ClampedZip2x9

Implements a clamped zip iterator combinator with 2 required iterators and 2 optional iterators.

ComponentNotFoundError

ExtraQueryHistory

When showing an entity in the history view, add this much history to it.

LatestAtComponentResults

Uncached results for a particular component when using a latest-at query.

LatestAtResults

Raw results for a latest-at query.

Promise

A Promise turns a source DataCell into a new DataCell with the helper of a PromiseResolver.

PromiseId

Uniquely identifies a Promise.

PromiseResolver

Resolves and keeps track of Promises.

RangeComponentResults

Uncached results for a particular component when using a range query.

RangeResults

Raw results for a range query.

RangeZip1x1

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip1x2

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip1x3

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip1x4

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip1x5

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip1x6

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip1x7

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip1x8

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip1x9

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip2x1

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip2x2

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip2x3

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip2x4

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip2x5

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip2x6

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip2x7

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip2x8

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

RangeZip2x9

Implements a range zip iterator combinator with 2 required iterators and 2 optional iterators.

VisibleHistory

Visible history bounds.

Enums

PromiseResult

The result of resolving a Promise through a PromiseResolver.

QueryError

VisibleHistoryBoundary

One of the boundaries of the visible history.

Traits

ToArchetype

Helper extension trait to convert query results into re_types_core::Archetypes.

Functions

clamped_zip_1x1

Returns a new ClampedZip1x1 iterator.

clamped_zip_1x2

Returns a new ClampedZip1x2 iterator.

clamped_zip_1x3

Returns a new ClampedZip1x3 iterator.

clamped_zip_1x4

Returns a new ClampedZip1x4 iterator.

clamped_zip_1x5

Returns a new ClampedZip1x5 iterator.

clamped_zip_1x6

Returns a new ClampedZip1x6 iterator.

clamped_zip_1x7

Returns a new ClampedZip1x7 iterator.

clamped_zip_1x8

Returns a new ClampedZip1x8 iterator.

clamped_zip_1x9

Returns a new ClampedZip1x9 iterator.

clamped_zip_2x1

Returns a new ClampedZip2x1 iterator.

clamped_zip_2x2

Returns a new ClampedZip2x2 iterator.

clamped_zip_2x3

Returns a new ClampedZip2x3 iterator.

clamped_zip_2x4

Returns a new ClampedZip2x4 iterator.

clamped_zip_2x5

Returns a new ClampedZip2x5 iterator.

clamped_zip_2x6

Returns a new ClampedZip2x6 iterator.

clamped_zip_2x7

Returns a new ClampedZip2x7 iterator.

clamped_zip_2x8

Returns a new ClampedZip2x8 iterator.

clamped_zip_2x9

Returns a new ClampedZip2x9 iterator.

latest_at

Queries for the given component_names using latest-at semantics.

range

Queries for the given component_names using range semantics.

range_zip_1x1

Returns a new RangeZip1x1 iterator.

range_zip_1x2

Returns a new RangeZip1x2 iterator.

range_zip_1x3

Returns a new RangeZip1x3 iterator.

range_zip_1x4

Returns a new RangeZip1x4 iterator.

range_zip_1x5

Returns a new RangeZip1x5 iterator.

range_zip_1x6

Returns a new RangeZip1x6 iterator.

range_zip_1x7

Returns a new RangeZip1x7 iterator.

range_zip_1x8

Returns a new RangeZip1x8 iterator.

range_zip_1x9

Returns a new RangeZip1x9 iterator.

range_zip_2x1

Returns a new RangeZip2x1 iterator.

range_zip_2x2

Returns a new RangeZip2x2 iterator.

range_zip_2x3

Returns a new RangeZip2x3 iterator.

range_zip_2x4

Returns a new RangeZip2x4 iterator.

range_zip_2x5

Returns a new RangeZip2x5 iterator.

range_zip_2x6

Returns a new RangeZip2x6 iterator.

range_zip_2x7

Returns a new RangeZip2x7 iterator.

range_zip_2x8

Returns a new RangeZip2x8 iterator.

range_zip_2x9

Returns a new RangeZip2x9 iterator.

Type Aliases

Result