Struct re_space_view::SpaceViewBlueprint

pub struct SpaceViewBlueprint {
    pub id: SpaceViewId,
    pub display_name: Option<String>,
    class_identifier: SpaceViewClassIdentifier,
    pub space_origin: EntityPath,
    pub contents: SpaceViewContents,
    pub visible: bool,
    pending_writes: Vec<DataRow>,
}

A view of a space.

Note: SpaceViewBlueprint doesn’t implement Clone because it stores an internal uuid used for identifying the path of its data in the blueprint store. It’s ambiguous whether the intent is for a clone to write to the same place.

If you want a new space view otherwise identical to an existing one, use re_viewport::ViewportBlueprint::duplicate_space_view.

Fields
id: SpaceViewIddisplay_name: Option<String>class_identifier: SpaceViewClassIdentifierspace_origin: EntityPath

The “anchor point” of this space view. The transform at this path forms the reference point for all scene->world transforms in this space view. I.e. the position of this entity path in space forms the origin of the coordinate system in this space view. Furthermore, this is the primary indicator for heuristics on what entities we show in this space view.

contents: SpaceViewContents

The content of this space view as defined by its queries.

visible: bool

True if this space view is visible in the UI.

pending_writes: Vec<DataRow>

Pending blueprint writes for nested components from duplicate.

Implementations

impl SpaceViewBlueprint

pub fn new( space_view_class: SpaceViewClassIdentifier, recommended: RecommendedSpaceView ) -> Self

Creates a new SpaceViewBlueprint with a single SpaceViewContents.

This SpaceViewBlueprint is ephemeral. If you want to make it permanent you must call Self::save_to_blueprint_store.

pub fn missing_name_placeholder(&self) -> String

Placeholder name displayed in the UI if the user hasn’t explicitly named the space view.

pub fn display_name_or_default(&self) -> ContentsName

Returns this space view’s display name

When returning ContentsName::Placeholder, the UI should display the resulting name using re_ui::LabelStyle::Unnamed.

pub fn try_from_db( id: SpaceViewId, blueprint_db: &EntityDb, query: &LatestAtQuery ) -> Option<Self>

Attempt to load a SpaceViewBlueprint from the blueprint store.

pub fn save_to_blueprint_store(&self, ctx: &ViewerContext<'_>)

Persist the entire SpaceViewBlueprint to the blueprint store.

This only needs to be called if the SpaceViewBlueprint was created with Self::new.

Otherwise, incremental calls to set_ functions will write just the necessary component update directly to the store.

pub fn duplicate( &self, store_context: &StoreContext<'_>, query: &LatestAtQuery ) -> Self

Creates a new SpaceViewBlueprint with the same contents, but a different SpaceViewId

Also duplicates all the queries in the space view.

pub fn clear(&self, ctx: &ViewerContext<'_>)

pub fn set_display_name(&self, ctx: &ViewerContext<'_>, name: Option<String>)

pub fn set_origin(&self, ctx: &ViewerContext<'_>, origin: &EntityPath)

pub fn set_visible(&self, ctx: &ViewerContext<'_>, visible: bool)

pub fn class_identifier(&self) -> &SpaceViewClassIdentifier

pub fn class<'a>( &self, space_view_class_registry: &'a SpaceViewClassRegistry ) -> &'a dyn SpaceViewClass

pub fn on_frame_start( &self, ctx: &ViewerContext<'_>, view_state: &mut dyn SpaceViewState, auto_properties: &mut EntityPropertyMap )

pub fn scene_ui( &self, view_state: &mut dyn SpaceViewState, ctx: &ViewerContext<'_>, ui: &mut Ui, query: &ViewQuery<'_>, system_output: SystemExecutionOutput )

pub fn entity_path(&self) -> EntityPath

pub fn legacy_properties( &self, blueprint: &EntityDb, blueprint_query: &LatestAtQuery ) -> EntityProperties

Legacy EntityProperties used by a hand ful of view properties that aren’t blueprint view properties yet.

pub fn save_legacy_properties( &self, ctx: &ViewerContext<'_>, props: EntityProperties )

pub fn query_range( &self, blueprint: &EntityDb, blueprint_query: &LatestAtQuery, active_timeline: &Timeline, space_view_class_registry: &SpaceViewClassRegistry ) -> QueryRange